Moved some code to new class files, renames, etc. PCDrv spits out gibberish and crashes psx
