+++ nops_proj.csproj
@@ -184,10 +184,12 @@
     <None Include="README.md" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="DataPort.cs" />
+    <Compile Include="DPSerial.cs" />
     <Compile Include="GDB.cs" />
     <Compile Include="NOTPSXSERIAL.CS" />
     <Compile Include="PCDrv.cs" />
-    <Compile Include="TCPBridge.cs" />
+    <Compile Include="DPTCPBridge.cs" />
     <Compile Include="TransferLogic.cs" />
     <Compile Include="Utils.cs" />
   </ItemGroup>
